requirements.txt
azure-identity==1.15.0
langchain==0.0.329
langchain-community==0.0.38
langchain-core==0.3.76
langchain-openai==0.3.33
langchain-text-splitters==0.0.2
langsmith==0.4.31
azure-search-documents==11.4.0b8
openai==1.109.1
python-dotenv==1.0.0
fastapi==0.108.0
uvicorn[standard]==0.25.0
tiktoken==0.11.0
